// database/prisma-thinkthank/generated/zod/inputTypeSchemas/TransactionAttachmentMinOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client'

import { z } from 'zod'
import { SortOrderSchema } from './SortOrderSchema'

export const TransactionAttachmentMinOrderByAggregateInputSchema: z.ZodType<Prisma.TransactionAttachmentMinOrderByAggregateInput> = z
    .object({
        id: z.lazy(() => SortOrderSchema).optional(),
        transactionId: z.lazy(() => SortOrderSchema).optional(),
        name: z.lazy(() => SortOrderSchema).optional(),
        description: z.lazy(() => SortOrderSchema).optional(),
        type: z.lazy(() => SortOrderSchema).optional(),
        category: z.lazy(() => SortOrderSchema).optional(),
        path: z.lazy(() => SortOrderSchema).optional(),
        size: z.lazy(() => SortOrderSchema).optional(),
        mimeType: z.lazy(() => SortOrderSchema).optional(),
        hash: z.lazy(() => SortOrderSchema).optional(),
        isConfidential: z.lazy(() => SortOrderSchema).optional(),
        status: z.lazy(() => SortOrderSchema).optional(),
        processingErrors: z.lazy(() => SortOrderSchema).optional(),
        ocrText: z.lazy(() => SortOrderSchema).optional(),
        createdAt: z.lazy(() => SortOrderSchema).optional(),
        updatedAt: z.lazy(() => SortOrderSchema).optional(),
        projectId: z.lazy(() => SortOrderSchema).optional()
    })
    .strict()

export default TransactionAttachmentMinOrderByAggregateInputSchema